+++ lib/gtksheet/gsheetmodel.h
@@ -51,8 +51,8 @@ typedef struct _GtkSheetCellBorder     GtkSheetCellBorder;
 
 struct _GtkSheetRange
 {
-  gint row0,col0; /* upper-left cell */
-  gint rowi,coli; /* lower-right cell */
+  glong row0,col0; /* upper-left cell */
+  glong rowi,coli; /* lower-right cell */
 };
 
 struct _GtkSheetCellBorder
@@ -75,20 +75,20 @@ struct _GSheetModelIface
 
   /* Signals */
   void         (* range_changed)    (GSheetModel *sheet_model,
-				     gint row0, gint col0,
-				     gint rowi, gint coli);
+				     glong row0, glong col0,
+				     glong rowi, glong coli);
 
   void         (* rows_inserted)    (GSheetModel *sheet_model,
-				     gint row, gint n_rows);
+				     glong row, glong n_rows);
 
   void         (* rows_deleted)     (GSheetModel *sheet_model,
-				     gint row, gint n_rows);
+				     glong row, glong n_rows);
 
   void         (* columns_inserted)    (GSheetModel *sheet_model,
-				     gint column, gint n_columns);
+				     glong column, glong n_columns);
 
   void         (* columns_deleted)     (GSheetModel *sheet_model,
-				     gint column, gint n_columns);
+				     glong column, glong n_columns);
 
 
 
@@ -96,36 +96,36 @@ struct _GSheetModelIface
   /* Virtual Table */
 
   gchar *      (* get_string)      (const GSheetModel *sheet_model,
-					       gint row, gint column);
+					       glong row, glong column);
 
   gboolean  (* set_string) (GSheetModel *sheet_model,
-			    const gchar *s, gint row, gint column);
+			    const gchar *s, glong row, glong column);
 
   gboolean  (* clear_datum) (GSheetModel *sheet_model,
-			     gint row, gint column);
+			     glong row, glong column);
 
-  gboolean (* is_visible) (const GSheetModel *sheet_model, gint row, gint column);
-  gboolean (* is_editable) (const GSheetModel *sheet_model, gint row, gint column);
+  gboolean (* is_visible) (const GSheetModel *sheet_model, glong row, glong column);
+  gboolean (* is_editable) (const GSheetModel *sheet_model, glong row, glong column);
 
   const GdkColor *  (* get_foreground) (const GSheetModel *sheet_model,
-				    gint row, gint column);
+				    glong row, glong column);
 
   const GdkColor *  (* get_background) (const GSheetModel *sheet_model,
-				    gint row, gint column);
+				    glong row, glong column);
 
   const GtkJustification *  (* get_justification) (const GSheetModel *sheet_model,
-						   gint row, gint column);
+						   glong row, glong column);
 
   const PangoFontDescription *  (* get_font_desc) (const GSheetModel *sheet_model,
-						   gint row, gint column);
+						   glong row, glong column);
 
   const GtkSheetCellBorder *  (* get_cell_border) (const GSheetModel *sheet_model,
-						   gint row, gint column);
+						   glong row, glong column);
 
 
-  gint (*get_column_count) (const GSheetModel *model);
+  glong (*get_column_count) (const GSheetModel *model);
 
-  gint (*get_row_count) (const GSheetModel *model);
+  glong (*get_row_count) (const GSheetModel *model);
 
 };
 
@@ -135,62 +135,62 @@ GType              g_sheet_model_get_type   (void) G_GNUC_CONST;
 
 
 inline  gchar * g_sheet_model_get_string (const GSheetModel *sheet_model,
-					       gint row, gint column);
+					       glong row, glong column);
 
 inline gboolean  g_sheet_model_set_string (GSheetModel *sheet_model,
 				      const gchar *s,
-				      gint row, gint column);
+				      glong row, glong column);
 
 inline gboolean g_sheet_model_datum_clear    (GSheetModel *sheet_model,
-					 gint row, gint column);
+					 glong row, glong column);
 
 
 inline void g_sheet_model_range_changed (GSheetModel *sheet_model,
-				    gint row0, gint col0,
-				    gint rowi, gint coli);
+				    glong row0, glong col0,
+				    glong rowi, glong coli);
 
 inline void g_sheet_model_rows_deleted (GSheetModel *sheet_model,
-				   gint row, gint n_rows);
+				   glong row, glong n_rows);
 
 inline void g_sheet_model_rows_inserted (GSheetModel *sheet_model,
-				    gint row, gint n_rows);
+				    glong row, glong n_rows);
 
 inline void g_sheet_model_columns_inserted (GSheetModel *sheet_model,
-					    gint column, gint n_columns);
+					    glong column, glong n_columns);
 
 inline void g_sheet_model_columns_deleted (GSheetModel *sheet_model,
-					   gint column, gint n_columns);
+					   glong column, glong n_columns);
 
 
 inline gboolean g_sheet_model_is_editable (const GSheetModel *model,
-				      gint row, gint column);
+				      glong row, glong column);
 
 inline gboolean g_sheet_model_is_visible
-                   (const GSheetModel *model, gint row, gint column);
+                   (const GSheetModel *model, glong row, glong column);
 
 
 inline const GdkColor *g_sheet_model_get_foreground
-                   (const GSheetModel *model, gint row, gint column);
+                   (const GSheetModel *model, glong row, glong column);
 
 inline const GdkColor *g_sheet_model_get_background
-                   (const GSheetModel *model, gint row, gint column);
+                   (const GSheetModel *model, glong row, glong column);
 
 
 inline const GtkJustification *g_sheet_model_get_justification
-                   (const GSheetModel *model, gint row, gint column);
+                   (const GSheetModel *model, glong row, glong column);
 
 
 inline const PangoFontDescription *g_sheet_model_get_font_desc
-                   (const GSheetModel *model, gint row, gint column);
+                   (const GSheetModel *model, glong row, glong column);
 
 inline const GtkSheetCellBorder * g_sheet_model_get_cell_border
-                   (const GSheetModel *model, gint row, gint column);
+                   (const GSheetModel *model, glong row, glong column);
 
 inline  gboolean g_sheet_model_free_strings (const GSheetModel *sheet_model);
 
-inline gint g_sheet_model_get_column_count(const GSheetModel *sheet_model);
+inline glong g_sheet_model_get_column_count (const GSheetModel *sheet_model);
 
-inline gint g_sheet_model_get_row_count(const GSheetModel *sheet_model);
+inline  gint g_sheet_model_get_row_count (const GSheetModel *sheet_model);
 
 G_END_DECLS
